+++ src/main/java/com/nasnav/dto/ReturnedItemList.java
@@ -4,12 +4,9 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
-
 @Data
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
 public class ReturnedItemList {
     private Long returnRequestItemId;
-    private BigDecimal receivedQuantity;
+    private Integer receivedQuantity;
 }
